Refactor: Persons class extracted to a dedicated file
